+++ R/importNI.R
@@ -15,7 +15,10 @@ importNI <- function(site = "bel0", year = 2018, data_type = "hourly",
                     data_type, "_NI_", year, ".rds")
     
     
-    aq_data <- map_df(files, readSummaryData, data_type = data_type, to_narrow = to_narrow)
+    aq_data <- map_df(files, readSummaryData, 
+                      data_type = data_type, 
+                      to_narrow = to_narrow,
+                      hc = FALSE)
     
     # add meta data?
     if (meta) {