+++ R/importSAQN.R
@@ -14,7 +14,10 @@ importSAQN <- function(site = "gla4", year = 2009, data_type = "hourly",
                     data_type, "_SCOT_", year, ".rds")
     
     
-    aq_data <- map_df(files, readSummaryData, data_type = data_type, to_narrow = to_narrow)
+    aq_data <- map_df(files, readSummaryData, 
+                      data_type = data_type, 
+                      to_narrow = to_narrow,
+                      hc = FALSE)
     
     # add meta data?
     if (meta) {