+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Do not return hydrocarbon measurements by default for annual and monthly summaries in `importAURN`.
 - fix issue with class of date-time object for importing some air quality data e.g. `importAURN`.
+- refine default grid lines in `windRose` and `pollutionRose`.
 
 # openair 2.9-0